+++ include/swift/Driver/Compilation.h
@@ -265,9 +265,6 @@ public:
   const bool OnlyOneDependencyFile;
 
 private:
-  /// Is the parser recording token hashes for each type body?
-  const bool EnableTypeFingerprints;
-
   /// Helpful for debugging, but slows down the driver. So, only turn on when
   /// needed.
   const bool VerifyFineGrainedDependencyGraphAfterEveryImport;
@@ -319,8 +316,6 @@ public:
               bool ShowDriverTimeCompilation = false,
               std::unique_ptr<UnifiedStatsReporter> Stats = nullptr,
               bool OnlyOneDependencyFile = false,
-              bool EnableTypeFingerprints =
-                LangOptions().EnableTypeFingerprints,
               bool VerifyFineGrainedDependencyGraphAfterEveryImport = false,
               bool EmitFineGrainedDependencyDotFileAfterEveryImport = false,
               bool FineGrainedDependenciesIncludeIntrafileOnes = false,
@@ -385,8 +380,6 @@ public:
     return EnableIncrementalBuild;
   }
   void disableIncrementalBuild(Twine why);
-
-  bool getEnableTypeFingerprints() const { return EnableTypeFingerprints; }
 
   bool getVerifyFineGrainedDependencyGraphAfterEveryImport() const {
     return VerifyFineGrainedDependencyGraphAfterEveryImport;
