+++ include/swift/Option/Options.td
@@ -139,14 +139,6 @@ def driver_always_rebuild_dependents :
   Flag<["-"], "driver-always-rebuild-dependents">, InternalDebugOpt,
   HelpText<"Always rebuild dependents of files that have been modified">;
 
-def enable_type_fingerprints :
-Flag<["-"], "enable-type-fingerprints">, Flags<[FrontendOption, HelpHidden]>,
-HelpText<"Enable per-nominal and extension body fingerprints">;
-
-def disable_type_fingerprints :
-Flag<["-"], "disable-type-fingerprints">, Flags<[FrontendOption, HelpHidden]>,
-HelpText<"Disable per-nominal and extension body fingerprints">;
-
 def enable_only_one_dependency_file :
 Flag<["-"], "enable-only-one-dependency-file">, Flags<[DoesNotAffectIncrementalBuild]>,
  HelpText<"Enables incremental build optimization that only produces one dependencies file">;
