+++ lib/Parse/ParseDecl.cpp
@@ -4551,8 +4551,13 @@ Parser::parseDeclList(SourceLoc LBLoc, SourceLoc &RBLoc, Diag<> ErrorDiag,
 
   // If we're hashing the type body separately, record the curly braces but
   // nothing inside for the interface hash.
+  //
+  // FIXME: There's no real reason code completion cannot also use this code
+  // path. But it seems to cause lazy parsing in contexts that the current
+  // implementation does not expect.
   Optional<llvm::SaveAndRestore<Optional<llvm::MD5>>> MemberHashingScope;
-  if (IDC->areTokensHashedForThisBodyInsteadOfInterfaceHash()) {
+  if (IDC->areTokensHashedForThisBodyInsteadOfInterfaceHash() &&
+      !L->isCodeCompletion()) {
     recordTokenHash("{");
     recordTokenHash("}");
     MemberHashingScope.emplace(CurrentTokenHash, llvm::MD5());
@@ -4587,7 +4592,7 @@ Parser::parseDeclList(SourceLoc LBLoc, SourceLoc &RBLoc, Diag<> ErrorDiag,
   if (RBLoc.isInvalid())
     hadError = true;
 
-  if (!Context.LangOpts.EnableTypeFingerprints)
+  if (L->isCodeCompletion())
     return std::make_pair(decls, None);
 
   llvm::MD5::MD5Result result;
