+++ unittests/Driver/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_swift_unittest(SwiftDriverTests
   FineGrainedDependencyGraphTests.cpp
   MockingFineGrainedDependencyGraphs.cpp
-  TypeBodyFingerprintsDependencyGraphTests.cpp
   UnitTestSourceFileDepGraphFactory.cpp
 )
